requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
# mysqlclient (Uncomment if you plan to use MariaDB/MySQL and install it)
